rename --srm option to --fm
