+++ mbed_host_tests/__init__.py
@@ -282,9 +282,11 @@ def init_host_test_cli_params():
                       dest='global_resource_mgr',
                       help='[Experimental] Global resource manager service module name, IP and port, example remote_client:10.2.123.43:3334')
 
-    parser.add_option('-s', '--srm',
-                      dest='simulator_resource_mgr',
-                      help='[Experimental] Simulator resource manager service module name, and a config file, example fastmodel_agent:DEFAULT')
+    parser.add_option('', '--fm',
+                      dest='fast_model_connection',
+                      metavar="CONFIG",
+                      default='DEFAULT',
+                      help='Fast Model connection, This option requires mbed-fastmodel-agent module installed, list CONFIGs via "mbedfm" ')
 
     parser.add_option('', '--run',
                       dest='run_binary',
